+++ src/main/java/org/opensearch/index/store/niofs/CryptoOutputStreamIndexOutput.java
@@ -9,76 +9,85 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Path;
 
-import javax.crypto.BadPaddingException;
-import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-
 import org.apache.lucene.store.OutputStreamIndexOutput;
+import org.opensearch.common.SuppressForbidden;
+import org.opensearch.index.store.cipher.OpenSslNativeCipher;
 
 /**
- * An IndexOutput implementation that encrypts data before writing
+ * An IndexOutput implementation that encrypts data before writing using native
+ * OpenSSL AES-CTR.
  *
  * @opensearch.internal
  */
-
+@SuppressForbidden(reason = "temporary bypass")
 public final class CryptoOutputStreamIndexOutput extends OutputStreamIndexOutput {
 
-    static final int CHUNK_SIZE = 8192;
+    private static final int CHUNK_SIZE = 8_192;
+    private static final int BUFFER_SIZE = 65_536;
 
     /**
-    * Creates a new CryptoIndexOutput
-    *
-    * @param name The name of the output
-    * @param path The path to write to
-    * @param os The output stream
-    * @param cipher The cipher to use for encryption
-    * @throws IOException If there is an I/O error
-    */
-    public CryptoOutputStreamIndexOutput(String name, Path path, OutputStream os, Cipher cipher) throws IOException {
-        super("FSIndexOutput(path=\"" + path + "\")", name, new EncryptedOutputStream(os, cipher), CHUNK_SIZE);
+     * Creates a new CryptoIndexOutput
+     *
+     * @param name The name of the output
+     * @param path The path to write to
+     * @param os The output stream
+     * @param key The AES key (must be 32 bytes for AES-256)
+     * @param iv The initialization vector (must be 16 bytes)
+     * @throws IOException If there is an I/O error
+     * @throws IllegalArgumentException If key or iv lengths are invalid
+     */
+    public CryptoOutputStreamIndexOutput(String name, Path path, OutputStream os, byte[] key, byte[] iv) throws IOException {
+        super("FSIndexOutput(path=\"" + path + "\")", name, new EncryptedOutputStream(os, key, iv), CHUNK_SIZE);
     }
 
     private static class EncryptedOutputStream extends FilterOutputStream {
-        private final Cipher cipher;
+
+        private final byte[] key;
+        private final byte[] iv;
         private final byte[] buffer;
         private int bufferPosition = 0;
-        private static final int BUFFER_SIZE = 65536; // Increased buffer size (64KB)
+        private long streamOffset = 0;
+        private boolean isClosed = false;
 
-        EncryptedOutputStream(OutputStream os, Cipher cipher) {
+        EncryptedOutputStream(OutputStream os, byte[] key, byte[] iv) {
             super(os);
-            this.cipher = cipher;
+            this.key = key;
+            this.iv = iv;
             this.buffer = new byte[BUFFER_SIZE];
         }
 
         @Override
         public void write(byte[] b, int offset, int length) throws IOException {
-            if (length >= BUFFER_SIZE) {
-                // For large writes, flush any buffered content first
-                flushBuffer();
+            checkClosed();
+            if (b == null) {
+                throw new NullPointerException("Input buffer cannot be null");
+            }
+            if (offset < 0 || length < 0 || offset + length > b.length) {
+                throw new IndexOutOfBoundsException("Invalid offset or length");
+            }
+            if (length == 0)
+                return;
 
-                // Process large chunks directly
+            if (length >= BUFFER_SIZE) {
+                flushBuffer();
                 processAndWrite(b, offset, length);
             } else if (bufferPosition + length > BUFFER_SIZE) {
-                // Buffer would overflow, flush first
                 flushBuffer();
                 System.arraycopy(b, offset, buffer, bufferPosition, length);
                 bufferPosition += length;
             } else {
-                // Add to buffer
                 System.arraycopy(b, offset, buffer, bufferPosition, length);
                 bufferPosition += length;
             }
         }
 
-        private void processAndWrite(byte[] data, int offset, int length) throws IOException {
-            try {
-                byte[] encrypted = cipher.update(data, offset, length);
-                if (encrypted != null && encrypted.length > 0) {
-                    out.write(encrypted);
-                }
-            } catch (IllegalStateException e) {
-                throw new IOException("Cipher update failed: " + e.getMessage(), e);
+        @Override
+        public void write(int b) throws IOException {
+            checkClosed();
+            if (bufferPosition >= BUFFER_SIZE) {
+                flushBuffer();
             }
+            buffer[bufferPosition++] = (byte) b;
         }
 
         private void flushBuffer() throws IOException {
@@ -88,34 +97,49 @@ public final class CryptoOutputStreamIndexOutput extends OutputStreamIndexOutput
             }
         }
 
-        @Override
-        public void write(int b) throws IOException {
-            if (bufferPosition >= BUFFER_SIZE) {
-                flushBuffer();
-            }
-            buffer[bufferPosition++] = (byte) b;
-        }
-
-        @Override
-        @SuppressWarnings("ConvertToTryWithResources")
-        public void close() throws IOException {
+        private void processAndWrite(byte[] data, int offset, int length) throws IOException {
             try {
-                flushBuffer();
-                byte[] finalBytes = cipher.doFinal();
-                if (finalBytes != null && finalBytes.length > 0) {
-                    out.write(finalBytes);
-                }
-            } catch (IllegalBlockSizeException | BadPaddingException e) {
-                throw new IOException("Cipher finalization failed", e);
-            } finally {
-                super.close();
+                byte[] encrypted = OpenSslNativeCipher.encrypt(key, iv, slice(data, offset, length), streamOffset);
+                out.write(encrypted);
+                streamOffset += length;
+            } catch (Throwable t) {
+                throw new IOException("Encryption failed at offset " + streamOffset, t);
             }
         }
 
+        private byte[] slice(byte[] data, int offset, int length) {
+            if (offset == 0 && length == data.length) {
+                return data;
+            }
+            byte[] sliced = new byte[length];
+            System.arraycopy(data, offset, sliced, 0, length);
+            return sliced;
+        }
+
         @Override
-        public void flush() throws IOException {
-            flushBuffer();
-            out.flush();
+        public void close() throws IOException {
+            IOException exception = null;
+
+            try {
+                checkClosed();
+                flushBuffer();
+                // Lucene writes footer here.
+                // this will also flush the buffer.
+                super.close();
+            } catch (IOException e) {
+                exception = e;
+            } finally {
+                isClosed = true;
+            }
+
+            if (exception != null)
+                throw exception;
+        }
+
+        private void checkClosed() throws IOException {
+            if (isClosed) {
+                throw new IOException("Outout stream is already closed, this is unusual");
+            }
         }
     }
 }
